Movie Recommender System
This project is a movie recommender system that clusters users based on the PCS (Pearson Correlation Similarity) measure to solve for the closest in likings users to estimate ratings given to movies by a user which is a Collaborative Filtering Methodology. The Engine then uses K-Means clustering to cluster movies of similar genre into groups for boosting the performance and accuracy of predictions. Also, we have made a basic terminal app which asks the user to rate certain movies and the recommends genre based on ratings as well as show to closest users thus predicting age, sex etc.